requirements.txt
mysql-connector-python
mock
unittest
